requirements.txt
psycopg2-binary==2.9.10
asgiref==3.8.1
Django==5.1.7
django-cors-headers==3.13.0  # Add this line
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
dj_database_url==2.3.0
dotenv==1.1.0
whitenoise==6.9.0